+++ core/Lucy/Search/ANDQuery.c
@@ -46,12 +46,12 @@ ANDQuery_init(ANDQuery *self, Vector *children) {
 String*
 ANDQuery_To_String_IMP(ANDQuery *self) {
     ANDQueryIVARS *const ivars = ANDQuery_IVARS(self);
-    uint32_t num_kids = Vec_Get_Size(ivars->children);
+    size_t num_kids = Vec_Get_Size(ivars->children);
     if (!num_kids) { return Str_new_from_trusted_utf8("()", 2); }
     else {
         CharBuf *buf = CB_new(0);
         CB_Cat_Trusted_Utf8(buf, "(", 1);
-        for (uint32_t i = 0; i < num_kids; i++) {
+        for (size_t i = 0; i < num_kids; i++) {
             String *kid_string = Obj_To_String(Vec_Fetch(ivars->children, i));
             CB_Cat(buf, kid_string);
             DECREF(kid_string);
@@ -108,7 +108,7 @@ Matcher*
 ANDCompiler_Make_Matcher_IMP(ANDCompiler *self, SegReader *reader,
                              bool need_score) {
     ANDCompilerIVARS *const ivars = ANDCompiler_IVARS(self);
-    uint32_t num_kids = Vec_Get_Size(ivars->children);
+    size_t num_kids = Vec_Get_Size(ivars->children);
 
     if (num_kids == 1) {
         Compiler *only_child = (Compiler*)Vec_Fetch(ivars->children, 0);
@@ -118,7 +118,7 @@ ANDCompiler_Make_Matcher_IMP(ANDCompiler *self, SegReader *reader,
         Vector *child_matchers = Vec_new(num_kids);
 
         // Add child matchers one by one.
-        for (uint32_t i = 0; i < num_kids; i++) {
+        for (size_t i = 0; i < num_kids; i++) {
             Compiler *child = (Compiler*)Vec_Fetch(ivars->children, i);
             Matcher *child_matcher
                 = Compiler_Make_Matcher(child, reader, need_score);
