+++ core/Lucy/Search/PhraseQuery.c
@@ -138,11 +138,11 @@ PhraseQuery_Equals_IMP(PhraseQuery *self, Obj *other) {
 String*
 PhraseQuery_To_String_IMP(PhraseQuery *self) {
     PhraseQueryIVARS *const ivars = PhraseQuery_IVARS(self);
-    uint32_t  num_terms = Vec_Get_Size(ivars->terms);
+    size_t  num_terms = Vec_Get_Size(ivars->terms);
     CharBuf  *buf       = CB_new(0);
     CB_Cat(buf, ivars->field);
     CB_Cat_Trusted_Utf8(buf, ":\"", 2);
-    for (uint32_t i = 0; i < num_terms; i++) {
+    for (size_t i = 0; i < num_terms; i++) {
         Obj    *term        = Vec_Fetch(ivars->terms, i);
         String *term_string = Obj_To_String(term);
         CB_Cat(buf, term_string);
@@ -300,7 +300,7 @@ PhraseCompiler_Make_Matcher_IMP(PhraseCompiler *self, SegReader *reader,
     PhraseQueryIVARS *const parent_ivars
         = PhraseQuery_IVARS((PhraseQuery*)ivars->parent);
     Vector *const      terms     = parent_ivars->terms;
-    uint32_t           num_terms = Vec_Get_Size(terms);
+    size_t             num_terms = Vec_Get_Size(terms);
 
     // Bail if there are no terms.
     if (!num_terms) { return NULL; }
@@ -322,7 +322,7 @@ PhraseCompiler_Make_Matcher_IMP(PhraseCompiler *self, SegReader *reader,
 
     // Look up each term.
     Vector  *plists = Vec_new(num_terms);
-    for (uint32_t i = 0; i < num_terms; i++) {
+    for (size_t i = 0; i < num_terms; i++) {
         Obj *term = Vec_Fetch(terms, i);
         PostingList *plist
             = PListReader_Posting_List(plist_reader, parent_ivars->field, term);
